+++ src/ansi.js
@@ -1,11 +1,12 @@
 "use strict";
 
-const caret                    = require("./caret");
-const erase                    = require("./erase");
-const scroll                   = require("./scroll");
-const {bg, fg, reset, style}   = require("./sgr");
-const {simplify}               = require("./simplify");
-const {sanitize, slice, strip} = require("./strip");
+const caret                            = require("./caret");
+const erase                            = require("./erase");
+const scroll                           = require("./scroll");
+const {bg, fg, reset, style}           = require("./sgr");
+const {simplify}                       = require("./simplify");
+const {slice}                          = require("./slice");
+const {sanitize, strip, visibleLength} = require("./strip");
 
 const ansi = {
 	caret,
@@ -18,6 +19,7 @@ const ansi = {
 	bg,
 
 	strip,
+	visibleLength,
 	sanitize,
 	slice,
 	simplify,